Restructured test to reduce file sizes and added complete test for PUBLISH_HEADING_LEVELS
